herb-ai-identifier/requirements.txt
streamlit==1.55.0
google-generativeai
duckduckgo-search
pillow
pandas